Adjusting PointNumSprite Function _1.

+++ AdditiveIKLib/CPP/mqomaterial.cpp
@@ -3971,7 +3971,7 @@ void CMQOMaterial::DrawCommon(
 		if (DXUTGetGlobalTimer()) {
 			m_cb[currentrefposindex].time.x = (float)DXUTGetGlobalTimer()->GetTime();
 		}
-		m_cb[currentrefposindex].time.y = (float)renderobj.pmodel->GetRefPosPointSize();
+		m_cb[currentrefposindex].time.y = renderobj.pmodel->GetRefPosPointSize();
 
 		MODELBOUND modelbb;
 		modelbb.Init();
@@ -4044,7 +4044,7 @@ void CMQOMaterial::DrawCommon(
 		if (DXUTGetGlobalTimer()) {
 			m_cb[currentrefposindex].time.x = (float)DXUTGetGlobalTimer()->GetTime();
 		}
-		m_cb[currentrefposindex].time.y = (float)renderobj.pmodel->GetRefPosPointSize();
+		m_cb[currentrefposindex].time.y = renderobj.pmodel->GetRefPosPointSize();
 
 		MODELBOUND modelbb;
 		modelbb.Init();
@@ -4135,7 +4135,7 @@ void CMQOMaterial::DrawCommon(
 		if (DXUTGetGlobalTimer()) {
 			m_cb[currentrefposindex].time.x = (float)DXUTGetGlobalTimer()->GetTime();
 		}
-		m_cb[currentrefposindex].time.y = (float)renderobj.pmodel->GetRefPosPointSize();
+		m_cb[currentrefposindex].time.y = renderobj.pmodel->GetRefPosPointSize();
 
 		MODELBOUND modelbb;
 		modelbb.Init();
@@ -4273,7 +4273,7 @@ void CMQOMaterial::InstancingDrawCommon(RenderContext* rc, myRenderer::RENDEROBJ
 		if (DXUTGetGlobalTimer()) {
 			m_cb[0].time.x = (float)DXUTGetGlobalTimer()->GetTime();
 		}
-		m_cb[0].time.y = (float)renderobj.pmodel->GetRefPosPointSize();
+		m_cb[0].time.y = renderobj.pmodel->GetRefPosPointSize();
 
 		MODELBOUND modelbb;
 		modelbb.Init();
@@ -4338,7 +4338,7 @@ void CMQOMaterial::InstancingDrawCommon(RenderContext* rc, myRenderer::RENDEROBJ
 		if (DXUTGetGlobalTimer()) {
 			m_cb[0].time.x = (float)DXUTGetGlobalTimer()->GetTime();
 		}
-		m_cb[0].time.y = (float)renderobj.pmodel->GetRefPosPointSize();
+		m_cb[0].time.y = renderobj.pmodel->GetRefPosPointSize();
 
 		MODELBOUND modelbb;
 		modelbb.Init();
@@ -4417,7 +4417,7 @@ void CMQOMaterial::InstancingDrawCommon(RenderContext* rc, myRenderer::RENDEROBJ
 		if (DXUTGetGlobalTimer()) {
 			m_cb[0].time.x = (float)DXUTGetGlobalTimer()->GetTime();
 		}
-		m_cb[0].time.y = (float)renderobj.pmodel->GetRefPosPointSize();
+		m_cb[0].time.y = renderobj.pmodel->GetRefPosPointSize();
 
 		MODELBOUND modelbb;
 		modelbb.Init();
